Use deleted_at instead of job to filter last active role
