+++ spec/domain/group/deleted_people_spec.rb
@@ -29,7 +29,6 @@ describe Group::DeletedPeople do
       context 'when group has people without role' do
         before do
           role.destroy
-          role.person.update(last_active_role: role)
         end
 
         it 'finds those people' do
@@ -37,7 +36,6 @@ describe Group::DeletedPeople do
         end
         it 'doesn\'t find people with new role' do
           Group::TopLayer::TopAdmin.create(person: person, group: group)
-          person.update(last_active_role: nil)
 
           expect(person.roles.count).to eq 1
           expect(Group::DeletedPeople.deleted_for(group).count).to eq 0
@@ -45,7 +43,6 @@ describe Group::DeletedPeople do
 
         it 'doesn\'t find people with role with future deletion date' do
           role = Group::TopLayer::TopAdmin.create(person: person, group: group)
-          person.update(last_active_role: nil)
           role.update!(deleted_at: 1.day.from_now)
 
           expect(person.roles.count).to eq 1
@@ -54,7 +51,6 @@ describe Group::DeletedPeople do
 
         it 'finds people from other group in same layer' do
           child_role.destroy
-          child_person.update(last_active_role: child_role)
           expect(Group::DeletedPeople.deleted_for(child_group)).to include child_person
         end
       end
@@ -77,7 +73,6 @@ describe Group::DeletedPeople do
       context 'when group has people without role' do
         before do
           role.destroy
-          role.person.update(last_active_role: role)
         end
 
         it 'finds those people' do
@@ -86,7 +81,6 @@ describe Group::DeletedPeople do
 
         it 'doesn\'t find people with new role' do
           Group::TopLayer::TopAdmin.create(person: person, group: group)
-          person.update(last_active_role: nil)
 
           expect(person.roles.count).to eq 1
           expect(Group::DeletedPeople.deleted_for_multiple(all_groups).count).to eq 0
@@ -95,7 +89,6 @@ describe Group::DeletedPeople do
         it 'doesn\'t find people with role with future deletion date' do
           role = Group::TopLayer::TopAdmin.create(person: person, group: group)
           role.update!(deleted_at: 1.day.from_now)
-          person.update(last_active_role: nil)
 
           expect(person.roles.count).to eq 1
           expect(Group::DeletedPeople.deleted_for_multiple(all_groups).count).to eq 0
@@ -103,7 +96,6 @@ describe Group::DeletedPeople do
 
         it 'finds people from other group in same layer' do
           child_role.destroy
-          child_person.update(last_active_role: child_role)
           expect(Group::DeletedPeople.deleted_for_multiple([child_group, sibling_group])).to include child_person
         end
       end
@@ -138,8 +130,6 @@ describe Group::DeletedPeople do
       role_top.destroy
       role_top.update_column(:deleted_at, 1.hour.ago)
       role_bottom.destroy
-      role_top.person.update(last_active_role: role_top)
-      role_bottom.person.update(last_active_role: role_bottom)
       expect(Group::DeletedPeople.deleted_for(group).count).to eq 0
     end
 
@@ -147,8 +137,6 @@ describe Group::DeletedPeople do
       role_bottom.destroy
       role_bottom.update_column(:deleted_at, 1.hour.ago)
       role_top.destroy
-      role_bottom.person.update(last_active_role: role_bottom)
-      role_top.person.update(last_active_role: role_top)
       expect(Group::DeletedPeople.deleted_for(group)).to include person
     end
 
@@ -156,7 +144,6 @@ describe Group::DeletedPeople do
       role_top.destroy
       role_top.update_column(:deleted_at, 1.hour.ago)
       role_bottom.destroy
-      person.update(last_active_role: role_bottom)
       expect(Group::DeletedPeople.deleted_for(bottom_group)).to include person
     end
 
@@ -176,7 +163,7 @@ describe Group::DeletedPeople do
       Fabricate(Group::BottomLayer::Leader.name, group: bottom_group, person: top2.person,
                 created_at: Time.zone.now - 9.months)
       top2.destroy
-      person.update(last_active_role: role_top)
+      top2.update!(deleted_at: del)
 
       expect(Group::DeletedPeople.deleted_for(group)).to match_array([person])
     end
