+++ spec/domain/search_strategies/sql_spec.rb
@@ -35,7 +35,6 @@ describe SearchStrategies::Sql do
     role.update(created_at: Time.now - 1.year)
     role.destroy
     @deleted_leader = role.person
-    @deleted_leader.update(last_active_role: role)
 
     role = Fabricate(Group::BottomGroup::Member.name.to_sym, group: groups(:bottom_group_one_one))
     role.update(created_at: Time.now - 1.year)
