+++ spec/features/group/deleted_people_controller_spec.rb
@@ -21,16 +21,14 @@ describe Group::DeletedPeopleController, js: true do
     let(:user) { people(:top_leader) }
 
     before do
-      del1 = Fabricate(Group::BottomLayer::Member.name.to_sym,
+      Fabricate(Group::BottomLayer::Member.name.to_sym,
                 group: groups(:bottom_layer_one),
                 created_at: 1.year.ago,
                 deleted_at: 1.month.ago)
-      del1.person.update(last_active_role: del1)
-      del2 = Fabricate(Group::BottomGroup::Leader.name.to_sym,
+      Fabricate(Group::BottomGroup::Leader.name.to_sym,
                 group: groups(:bottom_group_one_one_one),
                 created_at: 1.year.ago,
                 deleted_at: 1.month.ago)
-      del2.person.update(last_active_role: del2)
 
       sign_in(user)
       visit group_deleted_people_path(group_id: group.id)
